refactor: split mouse hook platform modules
