+++ core/mouse_hook_contract.py
@@ -1,0 +1,38 @@
+"""
+Structural contract exposed by platform mouse hook implementations.
+"""
+
+from typing import Any, Callable, Protocol, runtime_checkable
+
+
+@runtime_checkable
+class MouseHookLike(Protocol):
+    debug_mode: bool
+    invert_vscroll: bool
+    invert_hscroll: bool
+    divert_mode_shift: bool
+    divert_dpi_switch: bool
+    _hid_gesture: Any
+
+    def register(self, event_type: str, callback: Callable[[Any], None]) -> None: ...
+    def block(self, event_type: str) -> None: ...
+    def unblock(self, event_type: str) -> None: ...
+    def reset_bindings(self) -> None: ...
+    def configure_gestures(
+        self,
+        enabled: bool = False,
+        threshold: int = 50,
+        deadzone: int = 40,
+        timeout_ms: int = 3000,
+        cooldown_ms: int = 500,
+    ) -> None: ...
+    def set_connection_change_callback(self, cb: Callable[[bool], None]) -> None: ...
+    def set_debug_callback(self, callback: Callable[[str], None]) -> None: ...
+    def set_gesture_callback(self, callback: Callable[[Any], None]) -> None: ...
+    @property
+    def device_connected(self) -> bool: ...
+    @property
+    def connected_device(self) -> Any: ...
+    def dump_device_info(self) -> Any: ...
+    def start(self) -> bool: ...
+    def stop(self) -> None: ...
